AI Prompt  Chat/requirements.txt
Flask
huggingface_hub
python-dotenv
